refactor: String 확장파일추가
